test(stripe): split customer scenarios into customer.test.ts
Pulls 3 customer-lifecycle pre-describe tests (signup customer,
single customers.create call across signup+upgrade, email sync) and
4 sub-describes (getCustomerCreateParams, duplicate prevention,
user/org collision, search API fallback) out of stripe.test.ts. Wraps
the moved content in a "stripe customer" describe so the file has a
clear domain root. stripe.test.ts loses ~940 lines.
